requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2